pyproject.toml
[project]
name = "mcp-query"
version = "0.1.0"
description = "Local MCP server for secure database queries with Apple Keychain credentials"
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.0.0",
    "keyring>=25.0.0",
    "sqlparse>=0.5.0",
    "pymysql>=1.1.0",
    "psycopg2-binary>=2.9.0",
    "PyYAML>=6.0",
]

[project.scripts]
mcp-query = "mcp_query.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/mcp_query"]
